change to decimals

+++ strcalc.R
@@ -2,7 +2,7 @@ compute <- function(input_string) {
   # Perform simple arithmetic encoded in input string
   # e.g. '1 + 2' -> 3, or '1 - 2' -> -1
   values   <- unlist(strsplit(input_string, ' '))
-  num_0    <- as.integer(values[1])
+  num_0    <- as.numeric(values[1])
   operator <- values[2]
   num_1    <- as.integer(values[3])
   result   <- switch(operator,
